Modificaciones en el crud de facturacion

+++ database/migrations/2021_05_12_160454_create_facturas_table.php
@@ -21,7 +21,7 @@ class CreateFacturasTable extends Migration
             $table->string('apellidos',40);
             $table->string('telefono');
             $table->string('direccion');
-            $table->integer('valor');
+            $table->decimal('valor',8,3);
             $table->timestamps();
         });
     }
